move the rtc wrapper stuff to its own file
